3  -- API_BDD_carter_cash/API/requirements.txt
alembic==1.14.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.2.1
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.6.12
cryptography==42.0.5
databases==0.9.0
DateTime==5.4
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
Mako==1.3.8
MarkupSafe==2.1.5
orjson==3.9.15
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyodbc==5.1.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.36.3
tomli==2.2.1
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
zope.interface==6.2